Simplify optional easy task 1 wording
This Easy options task is really strangely worded 'Instead of displaying unordered list items, manage the amount of tasks in state and let each task display its number. Yes, you could also do that with a simple ordered list, but where's the fun in that? Try using state.'

A quick search in the Discord reveals people often ask what this means. It sounds like there are 2 parts to this task:
1. Display a sequential number in ascending order for each tasks 
2. [The confusing part] manage the amount of tasks in state? Does this mean limit the number a tasks a user can enter? Or keep track of the total number of tasks? This part is quite ambiguous